+++ client/app/menu/menu_item/menu_item_move/menu_item_move.h
@@ -5,7 +5,6 @@
 #include <vector>
 
 namespace client {
-
 template <typename T = float>
   requires std::is_arithmetic_v<T>
 class MenuItemMove : public common::MenuItem {
@@ -24,7 +23,7 @@ public:
   virtual std::string_view getName() const override;
 
   std::vector<T> getCoords() const;
-  static int getArgsCount();
+  static size_t getArgsCount();
 };
 } // namespace client
